+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * 【poi    】     优化ExcelReader，采用只读模式（pr#2204@Gitee）
 * 【poi    】     优化ExcelBase，将alias放入
 * 【poi    】     优化ExcelBase，将alias放入
+* 【core   】     改进StrUtil#startWith、endWith性能
 
 ### 🐞Bug修复
 * 【core   】     修复ObjectUtil.hasNull传入null返回true的问题（pr#555@Gitee）